Rename files with the services
