+++ packages/crypto-kraken-exchange-services/src/services/kucoin/kucoin-rest-v1/index.ts
@@ -1,2 +1,2 @@
-export { KuCoinRestV1 } from './kucoin-service';
+export { KuCoinRestV1 } from './kucoin-rest-v1';
 export { KuCoinExchangeCredentials } from './kucoin-exchange-credentials';
