+++ packages/crypto-kraken-exchange-services/src/services/yobit/yobit-rest-v3/index.ts
@@ -1,2 +1,2 @@
-export { YobitRestV3 } from './yobit-service';
+export { YobitRestV3 } from './yobit-rest-v3';
 export { YobitExchangeCredentials } from './yobit-exchange-credentials';
